DataSource Spring definition moved to openxava, but not working yet
