+++ invoicedemo/pending.txt
@@ -1,11 +1,14 @@
 Pending for Spring Boot migration (tmr):
 
 - Datasource
-    - Quitar context.xml, ¿poner mismo ejemplos en application.properties?
-    - Mover a OPenXava:
+
+    - Mover a OPenXava: TMR ME QUEDÉ POR AQUÍ: YA ESTÁ MOVIDO PERO AHORA NO FUNCIONA
         SpringDataSourceJndiFactory
         Código en InvoicedemoApplication
-    - ¿Cómo se configura db de producción?
+    - Quitar context.xml, ¿poner mismo ejemplos en application.properties?
+    - Paquete y nombre para para OpenXavaAutoConfiguration
+    - Paquete para SpringDataSourceJndiFactory
+    - ¿Cómo se configura db de producción? Hay formas estandar en Spring Boot ¿documentar?
     - ¿Quieres que lo dejemos así, o prefieres que avancemos también hacia
         la opción 3 (que OpenXava obtenga el DataSource/EntityManagerFactory
         directamente del contexto de Spring) como objetivo de OpenXava 8?
