+++ openxava/pom.xml
@@ -291,6 +291,12 @@
 			<version>4.1.0</version>
 			<scope>provided</scope>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-web</artifactId>
+			<version>4.1.0</version>
+			<scope>provided</scope>
+		</dependency>
 
 	</dependencies>
