+++ openxava/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,2 @@
 com.openxava.naviox.web.NaviOXAutoConfiguration
+org.openxava.web.OpenXavaAutoConfiguration
